+++ src/document.cpp
@@ -9,6 +9,10 @@
 #include <algorithm>
 
 namespace adm {
+  namespace detail {
+    template class OptionalParameter<Version>;
+  }  // namespace detail
+
   Document::Document() { idAssigner_.document(this); }
 
   std::shared_ptr<Document> Document::create() {