+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added support for silent audioTrackUid references with ID 0. See `AudioTrackUid::isSilent` and `AudioTrackUid::getSilent`.
+- Added `audioFormatExtended` `version` attribute; this should be set for BS.2076-2 compliance: `document->set(Version("ITU-R_BS.2076-2"));`.
 
 ### Changed
 - Decimal times are now written without trailing zeros past 5 decimal places. To interoperate with ADM parsers which don't support more than 5 digits, users should round times in the ADM document before writing.